NeighborlyAPI/requirements.txt
azure-functions==1.7.2
pymongo==3.12.0       
                      